+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 Changes:
 
   * Added ability to search samples, libraries, library aliquots, pools, and boxes by freezer
+  * Added sample requisition ID field
 
 # 0.2.195